- done: post operations service implementation; - respective controllers for post grud operations; - updated models in request, view and proto

+++ db/entity/post.go
@@ -4,7 +4,7 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 
 type Post struct {
 	Id     primitive.ObjectID `json:"id,omitempty" bson:"_id,omitempty"`
-	UserId int                `json:"user_id,omitempty"`
+	UserId int64              `json:"user_id,omitempty"`
 	Title  string             `json:"title,omitempty"`
 	Body   string             `json:"body,omitempty"`
 }
